client/requirements.txt
grpcio==1.27.2
scikit-learn==0.22.2.post1
pandas==1.0.3
Flask==1.1.1
grpcio-tools==1.27.2
protobuf==3.11.3
Flask==1.1.1
requests==2.23.0
streamlit==0.56.0